+++ footerrr.php
@@ -1,0 +1,51 @@
+<footer class="bg-light">
+    <hr>
+    <div class="container">
+        <div class="row mt-4">
+            <div class="col-sm-6">
+                <h4>Adresse</h4>
+                <ul>
+                    <li>8, rue Viguet - 1227 Les Acacias</li>
+                </ul>
+                <h4>Téléphone</h4>
+                <ul>
+                    <li>+ 41 (0)22 308 60 10</li>
+                </ul>
+                <h4>Horaires</h4>
+                <ul>
+                    <li>Du lundi au vendredi, de 9h à 17H</li>
+                </ul>
+            </div>
+            <div class="col-sm-6">
+                <h4>Contact</h4>
+                <iframe src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2761.8802012085266!2d6.12650291555611!3d46.19294257911629!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x478c7b2e1299827b%3A0xb6bbc30dd8e5306f!2sr%C3%A9alise%20-%20magasin%20d&#39;informatique%20d&#39;occasion!5e0!3m2!1sfr!2sch!4v1605783655516!5m2!1sfr!2sch"
+                        width="100%" height="300" frameborder="0" style="border:0;" allowfullscreen=""
+                        aria-hidden="false" tabindex="0"></iframe>
+            </div>
+        </div>
+    </div><!--/container-->
+
+    <div class="container">
+        <div class="row text-center">
+            <div class="col-sm-12">
+                <hr>
+                <p>&#169 par Super Mario Agency 2020</p>
+            </div>
+        </div>
+    </div><!--/container-->
+
+</footer>
+
+<!-- Optional JavaScript -->
+<!-- jQuery first, then Popper.js, then Bootstrap JS -->
+<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
+        integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
+        crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"
+        integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q"
+        crossorigin="anonymous"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"
+        integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
+        crossorigin="anonymous"></script>
+</body>
+</html>
