+++ index.php
@@ -1,148 +1,4 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-          integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Lato:wght@400;700&display=swap" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css2?family=Nunito:wght@400;500&display=swap" rel="stylesheet">
-    <title>Local Taste</title>
-
-    <!--CSS Integrer-->
-    <style>
-        body {
-            font-family: 'Nunito', sans-serif;
-            background: rgb(194, 222, 255)
-        }
-
-        .h1:hover {
-            text-decoration-style: none;
-        }
-
-        h1 {
-            font-size: 3.5em;
-        }
-
-        h2 {
-            margin-top: 1em;
-            margin-bottom: 1em;
-            color: #183050;
-            letter-spacing: 0.08em;
-            text-transform: uppercase;
-            font-family: Lato, sans-serif;
-        }
-
-        h3 {
-            color: white;
-            /* text-shadow: 1px 1px 4px rgb(0, 0, 0); */
-            background-color: #E53C3A;
-            font-size: 1.5em;
-        }
-
-        h3 + p {
-            color: white;
-        }
-
-        h5 {
-            font-weight: bold;
-            font-family: Lato, sans-serif;
-        }
-
-        .price-tag {
-            color: #E53C3A;
-            font-weight: 600;
-        }
-
-        .img-unique {
-            height: 250px;
-            object-fit: cover;
-            margin-top: 1em;
-        }
-
-        .logo-fleur {
-            height: 15vh;
-            display: inline-block;
-        }
-
-        .img-section:hover {
-            opacity: 80%;
-            transition: width 2s;
-        }
-
-        header {
-            background-color: #fff;
-        }
-
-        .nav-link {
-            font-size: 1.2em;
-        }
-
-        .hero-background {
-            background-image: url(assets/img/hero-riviera.jpg);
-            background-repeat: no-repeat;
-            background-size: 100%;
-            height: 330px;
-        }
-
-        .fromulaire-reservation {
-            margin-top: 2%;
-            margin-left: 5%;
-            margin-right: 5%;
-
-            padding: 23px;
-            background-color: white;
-            opacity: 0.8;
-        }
-    </style>
-</head>
-<body>
-
-<header>
-    <!-- Bannière réduction -->
-    <div class="banner alert-info text-white p-1" style="text-align: center; background-color: #344764;"><i
-            class="icon-gift icon-white"></i>Utilisez le code COVID et bénéficiez de -15% sur TOUS les voyages !
-    </div>
-
-
-    <!-- Barre de navigation avec logo -->
-    <div class="container">
-        <nav class="navbar navbar-light navbar-expand-lg">
-            <div class="row">
-                <a href="index.html" class="navbar-brand">
-                    <img src="assets/img/logo-fleur.png" alt="Edelweiss" class="float-left logo-fleur mr-2">
-                    <p class="align-middle">
-                    <h1 class="h1-home"><span style="color:rgb(194, 31, 31);">local</span> <span
-                            style="color:rgb(223, 84, 84)">taste</span></h1></p>
-                </a>
-            </div>
-
-            <button class="navbar-toggler" data-toggle="collapse" data-target=#navbarCollapse><span
-                    class="navbar-toggler-icon"></span></button>
-            <div class="collapse navbar-collapse" id="navbarCollapse">
-                <ul class="navbar-nav ml-auto">
-                    <li>
-                        <a href="index.html" class="nav-link">Accueil</a>
-                    </li>
-                    <li>
-                        <a href="nos_preferences.html" class="nav-link">Nos coups de coeur</a>
-                    </li>
-                    <li>
-                        <a href="notre-entreprise.html" class="nav-link">Notre entreprise</a>
-                    </li>
-                    <li>
-                        <a href="contact.html" class="nav-link">Contact</a>
-                    </li>
-                </ul>
-            </div>
-        </nav>
-    </div>
-    <hr>
-</header>
+<?php include_once("headerrr.php"); ?>
 
 <section class="hero">
     <div class="container-fluid hero-background">
@@ -349,54 +205,5 @@
         </div>
     </div>
 </div>
-<footer class="bg-light">
-    <hr>
-    <div class="container">
-        <div class="row mt-4">
-            <div class="col-sm-6">
-                <h4>Adresse</h4>
-                <ul>
-                    <li>8, rue Viguet - 1227 Les Acacias</li>
-                </ul>
-                <h4>Téléphone</h4>
-                <ul>
-                    <li>+ 41 (0)22 308 60 10</li>
-                </ul>
-                <h4>Horaires</h4>
-                <ul>
-                    <li>Du lundi au vendredi, de 9h à 17H</li>
-                </ul>
-            </div>
-            <div class="col-sm-6">
-                <h4>Contact</h4>
-                <iframe src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2761.8802012085266!2d6.12650291555611!3d46.19294257911629!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x478c7b2e1299827b%3A0xb6bbc30dd8e5306f!2sr%C3%A9alise%20-%20magasin%20d&#39;informatique%20d&#39;occasion!5e0!3m2!1sfr!2sch!4v1605783655516!5m2!1sfr!2sch"
-                        width="100%" height="300" frameborder="0" style="border:0;" allowfullscreen=""
-                        aria-hidden="false" tabindex="0"></iframe>
-            </div>
-        </div>
-    </div><!--/container-->
 
-    <div class="container">
-        <div class="row text-center">
-            <div class="col-sm-12">
-                <hr>
-                <p>&#169 par Super Mario Agency 2020</p>
-            </div>
-        </div>
-    </div><!--/container-->
-
-</footer>
-
-<!-- Optional JavaScript -->
-<!-- jQuery first, then Popper.js, then Bootstrap JS -->
-<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
-        integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
-        crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"
-        integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q"
-        crossorigin="anonymous"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"
-        integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
-        crossorigin="anonymous"></script>
-</body>
-</html>
+<?php include_once("footerrr.php"); ?>
