+++ messages.csv
@@ -1,0 +1,20 @@
+Nom, E-mail, Tél, Message
+IBRAHIMOVIC; iedin@gmail.com; 0763417341; 
+Nom, E-mail, Tél, Message
+IBRAHIMOVIC; iedin@gmail.com; 0763417341; 
+Nom, E-mail, Tél, Message
+IBRAHIMOVIC; iedin@gmail.com; 0763417341; 
+Nom, E-mail, Tél, Message
+Detrestresloing; ijijj@msn.copkm; 0763417341; 4444444444444
+Nom, E-mail, Téléphone, Message
+; iedin@gmail.com; 0763417341; fdghfg
+Nom, E-mail, Téléphone, Message
+kjhkhjk; 121212@ffff.hh; 1234567891; 66666666666666666
+Nom, E-mail, Téléphone, Message
+kjhkhjk; 121212@ffff.hh; 1234567891; 66666666666666666
+Nom, E-mail, Téléphone, Message
+kjhkhjk; 121212@ffff.hh; 1234567891; 66666666666666666
+Nom, E-mail, Téléphone, Message
+DEVIDÉ; OKOK@MSN.COM; 0763417341; FDFDF
+Nom, E-mail, Téléphone, Message
+lionel; liojnel@mokok.ddd; 0789554666; dddddddddddddddd
